% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_diffs}
\alias{check_diffs}
\title{check_diffs}
\usage{
check_diffs(DF_prepared, DF, ID_var, group_var, area_var, x_var, width_plot)
}
\arguments{
\item{DF_prepared}{output of prepare_data()}

\item{DF}{output of create_polygons()}

\item{ID_var}{ID variable}

\item{group_var}{group variable}

\item{area_var}{area variable}

\item{x_var}{x axis variable}

\item{width_plot}{reduce x for plotting by this ratio}
}
\value{
list
}
\description{
check_diffs
}
