% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_linearpackcircles.R
\name{check_linearpackcircles}
\alias{check_linearpackcircles}
\title{check_linearpackcircles}
\usage{
check_linearpackcircles(
  DF,
  ID_var = "ID",
  group_var = "group",
  area_var = "area",
  x_var = "x",
  separation_factor = 1,
  width_plot = 100,
  height_group = 10,
  CHECKS_plots = FALSE,
  random_seed = 12
)
}
\arguments{
\item{DF}{input dataframe}

\item{ID_var}{ID variable}

\item{group_var}{group variable}

\item{area_var}{area variable}

\item{x_var}{x axis variable}

\item{separation_factor}{how much separation between groups}

\item{width_plot}{reduce x for plotting by this ratio}

\item{height_group}{height of y axis for each group}

\item{random_seed}{random seed to use}
}
\value{

}
\description{
check_linearpackcircles
}
