% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_overlaps}
\alias{check_overlaps}
\title{check_overlaps}
\usage{
check_overlaps(DF_polygons, CHECKS_plots = FALSE)
}
\arguments{
\item{DF_polygons}{input dataframe (should be the output of create polygons)}

\item{CHECKS_plots}{Should we show plots}
}
\value{

}
\description{
check_overlaps
}
