% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{create_polygons}
\alias{create_polygons}
\title{create_polygons}
\usage{
create_polygons(DF, group_var)
}
\arguments{
\item{DF}{input dataframe}

\item{group_var}{group variable}
}
\value{
DF
}
\description{
create_polygons
}
