% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearpackcircles.R
\name{linearpackcircles}
\alias{linearpackcircles}
\title{linearpackcircles}
\usage{
linearpackcircles(
  DF,
  ID_var = "ID",
  group_var = "group",
  area_var = "area",
  x_var = "x",
  separation_factor = 1,
  ratio_reduction_area = 600,
  ratio_reduction_x = 50,
  height_y = 1,
  max_overlaps = 8,
  title_str = "Packing circles visualization",
  subtitle_str = NULL,
  x_str = x_var,
  caption_str = "",
  size_text = 3,
  label_circles = TRUE,
  save_plot = FALSE,
  ...
)
}
\arguments{
\item{DF}{input dataframe}

\item{ID_var}{ID variable}

\item{group_var}{group variable}

\item{area_var}{area variable}

\item{x_var}{x axis variable}

\item{separation_factor}{how much separation between groups}

\item{ratio_reduction_area}{reduce area for plotting by this ratio}

\item{ratio_reduction_x}{reduce x for plotting by this ratio}

\item{height_y}{height of y axis for each group}

\item{max_overlaps}{overlaps in geom_text_repel}

\item{title_str}{title plot}

\item{subtitle_str}{subtitle plot}

\item{x_str}{x axis label plot}

\item{caption_str}{caption plot}

\item{size_text}{size text labels}

\item{label_circles}{Should we draw labels for the circles}

\item{save_plot}{Should we save the plot}

\item{...}{other arguments}
}
\value{

}
\description{
linearpackcircles
}
