% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{mult_format}
\alias{mult_format}
\title{mult_format
Return x axis labels to original values (we use ratio_reduction_x above to be able to perform calculations)}
\usage{
mult_format(ratio_reduction_x, width_plot)
}
\arguments{
\item{ratio_reduction_x}{automatically calculated parameter}

\item{width_plot}{reduce x for plotting by this ratio}
}
\value{
function
}
\description{
mult_format
Return x axis labels to original values (we use ratio_reduction_x above to be able to perform calculations)
}
