% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{prepare_data}
\alias{prepare_data}
\title{prepare_data}
\usage{
prepare_data(
  DF,
  ID_var,
  group_var,
  area_var,
  x_var,
  ratio_reduction_area,
  ratio_reduction_x,
  height_y
)
}
\arguments{
\item{DF}{input dataframe}

\item{ID_var}{ID variable}

\item{group_var}{group variable}

\item{area_var}{area variable}

\item{x_var}{x axis variable}

\item{ratio_reduction_area}{reduce area for plotting by this ratio}

\item{ratio_reduction_x}{reduce x for plotting by this ratio}

\item{height_y}{height of y axis for each group}
}
\value{

}
\description{
prepare_data
}
