% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_diffs}
\alias{check_diffs}
\title{check_diffs}
\usage{
check_diffs(
  DF_prepared,
  DF,
  check_var,
  ID_var,
  group_var,
  area_var,
  x_var,
  width_plot
)
}
\arguments{
\item{DF}{data2}

\item{check_var}{x variable}

\item{ID_var}{ID variable}

\item{group_var}{group variable}

\item{ALL_data}{data1}

\item{ratio_reduction_x}{ratio_reduction_x}
}
\value{

}
\description{
check_diffs
}
