% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{create_plot}
\alias{create_plot}
\title{create_plot}
\usage{
create_plot(
  DF_prepared,
  DF,
  ID_var,
  group_var,
  area_var,
  x_var,
  separation_factor = 5,
  width_plot = 100,
  label_circles = FALSE,
  max_overlaps = 5,
  size_text = 3,
  highlight_ID = NULL
)
}
\arguments{
\item{DF_prepared}{output of prepare_data()}

\item{DF}{output of create_polygons()}

\item{ID_var}{ID variable}

\item{group_var}{group variable}

\item{area_var}{area variable}

\item{x_var}{x axis variable}

\item{separation_factor}{how much separation between groups}

\item{width_plot}{reduce x for plotting by this ratio}

\item{label_circles}{Should we draw labels for the circles}

\item{max_overlaps}{overlaps in geom_text_repel}

\item{size_text}{size text labels}

\item{highlight_ID}{ID's to highlight in plot}
}
\value{
ggplot object
}
\description{
create_plot
}
