% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{prepare_data}
\alias{prepare_data}
\title{prepare_data}
\usage{
prepare_data(
  DF,
  ID_var,
  group_var,
  area_var,
  x_var,
  width_plot = 100,
  height_group = 10,
  area_multiplier = 1
)
}
\arguments{
\item{DF}{input dataframe}

\item{ID_var}{ID variable}

\item{group_var}{group variable}

\item{area_var}{area variable}

\item{x_var}{x axis variable}

\item{width_plot}{reduce x for plotting by this ratio}

\item{height_group}{height of y axis for each group}

\item{area_multiplier}{multiply area size by this}
}
\value{
DF
}
\description{
prepare_data
}
